Added functionality
Now, when the date of an event is modified, all the users who are going to participate to it are notified. If they have already accepted an invitation for 
another event and the two dates of the events overlaps, the user is notified about it and it is explained to him how he has to proceed in order to 
accept the invitation of the delayed event.