Fix of create event
Now the event is created correctly, added some exception in case of error in the creation.
AddInvitation now control if the email exist, if the user you want to invite is the creator or is already invited.
Now if someone call the addInvitation page, there's some control if the user that call this page is the creator of the event or not
Some little modification to importExport to adapt it to the new version of the createEvent method